Feat: move value and transformer to general tab
