+++ Makefile
@@ -117,7 +117,7 @@ ifeq "$(DISABLE_E2E_TESTS)" "true"
 else
 	$(MAKE) test-e2e-start-dockers
 	npx cypress install
-	./bin/wait-for -t 30 localhost:3000 -- npx cypress run --browser chrome || (\
+	./bin/wait-for -t 30 localhost:3000 -- npx cypress open --browser chrome || (\
 		$(MAKE) test-e2e-stop-dockers && \
 		exit 1)
 endif
