+++ src/app/custom/translations.tsv
@@ -187,8 +187,6 @@
 "after_field"	"After %{field}"	"Après %{field}"
 "first_position"	"At first position"	"En première position"
 "position"	"Position"	"Position"
-"field_wizard_tab_value"	"Value"	"Valeur"
-"field_wizard_tab_tranforms"	"Transformation"	"Transformation"
 "field_wizard_tab_identity"	"General"	"Général"
 "field_wizard_tab_display"	"Display"	"Affichage"
 "field_wizard_tab_search"	"Annotations"	"Annotations"
