+++ src/app/js/admin/preview/field/index.js
@@ -5,6 +5,7 @@ import {
     ARRAY_INSERT as REDUX_FORM_ARRAY_INSERT,
     ARRAY_REMOVE as REDUX_FORM_ARRAY_REMOVE,
     ARRAY_MOVE as REDUX_FORM_ARRAY_MOVE,
+    ARRAY_PUSH as REDUX_FORM_ARRAY_PUSH,
     REGISTER_FIELD as REDUX_FORM_REGISTER_FIELD,
     UNREGISTER_FIELD as REDUX_FORM_UNREGISTER_FIELD,
 } from 'redux-form/lib/actionTypes';
@@ -31,6 +32,7 @@ export default handleActions(
             REDUX_FORM_ARRAY_INSERT,
             REDUX_FORM_ARRAY_REMOVE,
             REDUX_FORM_ARRAY_MOVE,
+            REDUX_FORM_ARRAY_PUSH,
             REDUX_FORM_REGISTER_FIELD,
             REDUX_FORM_UNREGISTER_FIELD,
         )]: state => ({ ...state, isComputing: true }),
