+++ src/app/js/fields/wizard/TabSemantics.js
@@ -6,12 +6,14 @@ import { field as fieldPropTypes } from '../../propTypes';
 import FieldComposedOf from '../FieldComposedOf';
 import FieldAnnotation from '../FieldAnnotation';
 import FieldSchemeInput from '../FieldSchemeInput';
+import FieldLanguageInput from '../FieldLanguageInput';
 
-export const TabSemanticsComponent = ({ fields }) => {
+export const TabSemanticsComponent = ({ fields, field }) => {
     const { filter } = useParams();
     return (
         <>
             <FieldSchemeInput />
+            <FieldLanguageInput field={field} />
             <FieldAnnotation fields={fields} scope={filter} />
             <FieldComposedOf fields={fields} />
         </>
@@ -20,6 +22,7 @@ export const TabSemanticsComponent = ({ fields }) => {
 
 TabSemanticsComponent.propTypes = {
     fields: PropTypes.arrayOf(fieldPropTypes).isRequired,
+    field: fieldPropTypes.isRequired,
 };
 
 export default TabSemanticsComponent;
