+++ src/app/js/fields/wizard/ValuePreview.js
@@ -50,7 +50,9 @@ const ValuePreview = ({ lines, editedField, p: polyglot }) => {
                                     WebkitBoxOrient: 'vertical',
                                 }}
                             >
-                                {line[editedField.name]}
+                                {typeof line[editedField.name] === 'object'
+                                    ? JSON.stringify(line[editedField.name])
+                                    : line[editedField.name]}
                             </Typography>
                         </Box>
                     ))}
